requirements.txt
streamlit
PyPDF2
langchain
langchain_google_genai
langchain_community
python-dotenv
